# services/app_api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RowDetail": {
        "description": "``GET /api/rows/{row_key}``: the row with every mirror column, overrides, status log and siblings.",
        "properties": {
          "adjusted_need_by_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted Need By Date"
          },
          "adjusted_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted Reason Code"
          },
          "air_gap": {
            "title": "Air Gap",
            "type": "boolean"
          },
          "air_gap_hours": {
            "title": "Air Gap Hours",
            "type": "integer"
          },
          "batch_no": {
            "title": "Batch No",
            "type": "string"
          },
          "campaign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ChangeControlOut"
            },
            "title": "Changes",
            "type": "array"
          },
          "coa_release_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coa Release Reason"
          },
          "current_overrides": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OverrideOut"
            },
            "title": "Current Overrides",
            "type": "object"
          },
          "current_stage_entry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage Entry Date"
          },
          "days_in_stage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days In Stage"
          },
          "deviation_light": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviation Light"
          },
          "deviations": {
            "items": {
              "$ref": "#/components/schemas/DeviationOut"
            },
            "title": "Deviations",
            "type": "array"
          },
          "expedite": {
            "title": "Expedite",
            "type": "boolean"
          },
          "facts": {
            "additionalProperties": true,
            "title": "Facts",
            "type": "object"
          },
          "flags": {
            "$ref": "#/components/schemas/FlagsOut"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "gr_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gr Date"
          },
          "inbound_check": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InboundCheckOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "inbound_light": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inbound Light"
          },
          "inspection_lot_no": {
            "title": "Inspection Lot No",
            "type": "string"
          },
          "late": {
            "title": "Late",
            "type": "boolean"
          },
          "latest_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestStatusOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "lims_approved_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lims Approved Date"
          },
          "lims_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lims Status"
          },
          "location_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Type"
          },
          "lot_type": {
            "title": "Lot Type",
            "type": "string"
          },
          "manual_hold_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Hold Reason"
          },
          "material_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Class"
          },
          "material_desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Desc"
          },
          "material_no": {
            "title": "Material No",
            "type": "string"
          },
          "molecule_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Molecule Type"
          },
          "next_inspection_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Inspection Date"
          },
          "operative_need_by": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operative Need By"
          },
          "override_history": {
            "items": {
              "$ref": "#/components/schemas/OverrideOut"
            },
            "title": "Override History",
            "type": "array"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanOut"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "qc_testing_entry": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qc Testing Entry"
          },
          "row_key": {
            "title": "Row Key",
            "type": "string"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/SampleOut"
            },
            "title": "Samples",
            "type": "array"
          },
          "siblings": {
            "items": {
              "$ref": "#/components/schemas/RowOut"
            },
            "title": "Siblings",
            "type": "array"
          },
          "stage_key": {
            "title": "Stage Key",
            "type": "string"
          },
          "stage_label": {
            "title": "Stage Label",
            "type": "string"
          },
          "stage_rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Rule Id"
          },
          "status_log": {
            "items": {
              "$ref": "#/components/schemas/StatusLogOut"
            },
            "title": "Status Log",
            "type": "array"
          },
          "status_log_count": {
            "title": "Status Log Count",
            "type": "integer"
          },
          "storage_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Location"
          },
          "supplier_batch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Batch"
          },
          "supplier_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Name"
          },
          "system_need_by_locked": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Need By Locked"
          },
          "ud_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ud Code"
          },
          "ud_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ud Date"
          }
        },
        "required": [
          "row_key",
          "material_no",
          "material_desc",
          "material_class",
          "molecule_type",
          "supplier_name",
          "supplier_batch",
          "batch_no",
          "inspection_lot_no",
          "lot_type",
          "campaign",
          "storage_location",
          "location_type",
          "stage_key",
          "stage_label",
          "stage_rule_id",
          "current_stage_entry_date",
          "gr_date",
          "qc_testing_entry",
          "lims_approved_date",
          "ud_date",
          "next_inspection_date",
          "lims_status",
          "ud_code",
          "system_need_by_locked",
          "adjusted_need_by_date",
          "adjusted_reason_code",
          "operative_need_by",
          "expedite",
          "latest_status",
          "plan",
          "air_gap",
          "air_gap_hours",
          "late",
          "days_in_stage",
          "deviation_light",
          "inbound_light",
          "flags",
          "status_log_count",
          "sample_count",
          "manual_hold_reason",
          "coa_release_reason",
          "freshness",
          "facts",
          "current_overrides",
          "override_history",
          "status_log",
          "deviations",
          "inbound_check",
          "changes",
          "samples",
          "siblings"
        ],
        "title": "RowDetail",
        "type": "object"
      },
...
      "StatusLogPage": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/StatusLogOut"
            },
            "title": "Entries",
            "type": "array"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusLogOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "count",
          "latest",
          "entries"
        ],
        "title": "StatusLogPage",
        "type": "object"
      },
...
      "ToggleIn": {
        "description": "Body of Place/Release Hold and Release on COA/Undo: the reason is required either way (OQ-103).",
        "properties": {
          "on": {
            "title": "On",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "on",
          "reason"
        ],
        "title": "ToggleIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "R2R Intelligence application API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rows/{row_key}": {
      "get": {
        "operationId": "get_row_api_rows__row_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Row"
      }
    },
    "/api/rows/{row_key}/coa-release": {
      "post": {
        "operationId": "post_coa_release_api_rows__row_key__coa_release_post",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Coa Release"
      }
    },
...
    "/api/rows/{row_key}/hold": {
      "post": {
        "operationId": "post_hold_api_rows__row_key__hold_post",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Hold"
      }
    },
    "/api/rows/{row_key}/need-by": {
      "put": {
        "operationId": "put_need_by_api_rows__row_key__need_by_put",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NeedByIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Need By"
      }
    },
    "/api/rows/{row_key}/need-by/preview": {
      "post": {
        "description": "Recompute the row with the proposed override, through the composition a read uses. Writes nothing.",
        "operationId": "preview_need_by_api_rows__row_key__need_by_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NeedByIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Need By"
      }
    },
    "/api/rows/{row_key}/status-log": {
      "get": {
        "operationId": "get_status_log_api_rows__row_key__status_log_get",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusLogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Status Log"
      },
      "post": {
        "operationId": "post_status_log_api_rows__row_key__status_log_post",
        "parameters": [
          {
            "in": "path",
            "name": "row_key",
            "required": true,
            "schema": {
              "title": "Row Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-demo-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusLogIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusLogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Status Log"
      }
    },
...
  }
}